fix: Resolve mcp dev connection issues and move VisualState
- Moves VisualState class to its own file (omnimcp/visual_state.py).
- Renames omnimcp/omnimcp.py to omnimcp/mcp_server.py.
- Refactors mcp_server.py to instantiate FastMCP at module level, use mcp.run(), and require OMNIPARSER_URL via config, fixing connection errors with `mcp dev`.
- Updates imports and README architecture section for file changes.
- Modifies MCP tools (click_element, type_text) to remove second state update/verification to mitigate timeouts (though underlying performance issues remain).
