+++ pyproject.toml
@@ -52,6 +52,7 @@ test = [
     "pytest-mock>=3.10.0",
     "pytest-asyncio>=0.23.5",
     "ruff>=0.11.2",
+    "mcp[cli]",
 ]
 
 # Add Ruff configuration if you want to manage it here
